Core/Misc: some fixes.

+++ src/server/game/Loot/LootMgr.cpp
@@ -1390,7 +1390,7 @@ bool LootTemplate::addConditionItem(Condition* cond)
             {
                 for (LootStoreItemList::iterator i = itemList->begin(); i != itemList->end(); ++i)
                 {
-                    if ((*i).itemid == cond->SourceEntry)
+                    if ((*i).itemid == uint32(cond->SourceEntry))
                     {
                         (*i).conditions.push_back(cond);
                         return true;
@@ -1402,7 +1402,7 @@ bool LootTemplate::addConditionItem(Condition* cond)
             {
                 for (LootStoreItemList::iterator i = itemList->begin(); i != itemList->end(); ++i)
                 {
-                    if ((*i).itemid == cond->SourceEntry)
+                    if ((*i).itemid == uint32(cond->SourceEntry))
                     {
                         (*i).conditions.push_back(cond);
                         return true;
